Create explain.rs file, use libterm
